:construction: split utils.js
